refactor: merge 로직을 useMergedProducts hook으로 이동
